+++ app/src/main/java/com/mobithink/carbon/driving/DrivingEventDialogFragment.java
@@ -58,8 +58,8 @@ public class DrivingEventDialogFragment extends DialogFragment implements Locati
                 eventDTO = new EventDTO();
                 eventDTO.setEventName(eventType.get(position));
                 eventDTO.setStartTime(System.currentTimeMillis());
-                eventDTO.setGpsLat((long) latitude);
-                eventDTO.setGpsLong((long) longitude);
+                eventDTO.setGpsLat(latitude);
+                eventDTO.setGpsLong(longitude);
                 eventDTO.setId(DatabaseManager.getInstance().createNewEvent(CarbonApplicationManager.getInstance().getCurrentTripId(), null, eventDTO));
                 mListener.onEventSelected(eventDTO);
                 dismiss();
